+++ lavafume/main.lua
@@ -1,356 +1,1 @@
-mod_name, version = 'Lavafume', 1.08
-gh_script, gh_ui  = 'https://raw.githubusercontent.com/iammoonman/cera-roe/main/lavafume/main.lua', 'https://raw.githubusercontent.com/iammoonman/cera-roe/main/lavafume/ui.xml'
-IDToColor         = { ['k'] = 'Pink', ['w'] = 'White', ['b'] = 'Brown', ['r'] = 'Red', ['o'] = 'Orange', ['y'] = 'Yellow', ['g'] = 'Green', ['t'] = 'Teal', ['u'] = 'Blue', ['p'] = 'Purple' }
-mnrcState         = { ['k'] = '', ['w'] = '', ['b'] = '', ['r'] = '', ['o'] = '', ['y'] = '', ['g'] = '', ['t'] = '', ['u'] = '', ['p'] = '' }
-dyntState         = { ['k'] = '', ['w'] = '', ['b'] = '', ['r'] = '', ['o'] = '', ['y'] = '', ['g'] = '', ['t'] = '', ['u'] = '', ['p'] = '' }
-dayColor          = 'Yellow'
-nightColor        = 'Red'
-noColor           = ''
-
-function onLoad(s)
-    self.interactable = false;
-    WebRequest.get(gh_script, self, 'GetFreshScript')
-    if s ~= '' then
-        local state = JSON.decode(s)
-        mnrcState = state.mnrcState
-        dyntState = state.dyntState
-    end
-end
-
-function onSave()
-    return JSON.encode({mnrcState = mnrcState, dyntState = dyntState})
-end
-
-function GetFreshScript(wr)
-    if wr == nil then return end
-    local v = wr.text:match('mod_name, version = \'Lavafume\', (%d+%p%d+)')
-    log('GITHUB Version ' .. v)
-    if v then v = tonumber(v) else v = version end
-    if version < v then
-        self.setLuaScript(wr.text)
-        self.reload()
-    end
-    WebRequest.get(gh_ui, self, 'GetFreshXML')
-end
-
-function GetFreshXML(wr)
-    if wr == nil then return end
-    self.UI.setXml(wr.text)
-end
-
-function ping(player, _, button_id)
-    local color = IDToColor[button_id:match('_(%l)')]
-    for i, p in ipairs(Player.getPlayers()) do
-        if p.host then p.broadcast(color .. ' needs assistance.') end
-    end
-end
-
-function untap(player, _, button_id)
-    local color = IDToColor[button_id:match('_(%l)')]
-    local guid = '5a1314'
-    local plane_rotation = 270
-    if color == 'Pink' then guid = '6b2479' plane_rotation = 180 end
-    if color == 'White' then guid = 'ed8834' plane_rotation = 180 end
-    if color == 'Brown' then guid = '1451b7' plane_rotation = 180 end
-    if color == 'Red' then guid = '2c271a' plane_rotation = 180 end
-    if color == 'Orange' then guid = '00a854' plane_rotation = 180 end
-    if color == 'Yellow' then guid = '6c87b2' plane_rotation = 0 end
-    if color == 'Green' then guid = 'fd020e' plane_rotation = 0 end
-    if color == 'Teal' then guid = '90ea3b' plane_rotation = 0 end
-    if color == 'Blue' then guid = '6180e9' plane_rotation = 0 end
-    if color == 'Purple' then guid = '7058c4' plane_rotation = 0 end
-    local zone = getObjectFromGUID(guid)
-    for _, occupyingObject in ipairs(zone.getObjects()) do
-        if occupyingObject.type == "Card" then
-            local rotation = occupyingObject.getRotation()
-            local flip_angle = rotation.z
-            occupyingObject.setRotationSmooth({ x = 0, y = plane_rotation, z = flip_angle })
-        end
-    end
-end
-
-function g1(player, _, button_id)
-    local button_a = 'g1_k'
-    local button_b = 'g1_p'
-    local color = IDToColor[button_id:match('_(%l)')]
-    if button_id == 'g1_p' or button_id == 'g1_k' then button_a = 'g1_k' button_b = 'g1_p'
-    elseif button_id == 'g1_w' or button_id == 'g1_u' then button_a = 'g1_w' button_b = 'g1_u'
-    elseif button_id == 'g1_b' or button_id == 'g1_t' then button_a = 'g1_b' button_b = 'g1_t'
-    elseif button_id == 'g1_r' or button_id == 'g1_g' then button_a = 'g1_r' button_b = 'g1_g'
-    elseif button_id == 'g1_o' or button_id == 'g1_y' then button_a = 'g1_o' button_b = 'g1_y'
-    end
-    if self.UI.getAttribute(button_id, 'outline') ~= color then
-        self.UI.setAttribute(button_a, 'outline', color)
-        self.UI.setAttribute(button_b, 'outline', color)
-    else
-        self.UI.setAttribute(button_a, 'outline', noColor)
-        self.UI.setAttribute(button_b, 'outline', noColor)
-    end
-end
-
-function g2(player, _, button_id)
-    local button_a = 'g2_p'
-    local button_b = 'g2_k'
-    local color = IDToColor[button_id:match('_(%l)')]
-    if button_id == 'g2_p' or button_id == 'g2_k' then button_a = 'g2_k' button_b = 'g2_p'
-    elseif button_id == 'g2_w' or button_id == 'g2_u' then button_a = 'g2_w' button_b = 'g2_u'
-    elseif button_id == 'g2_b' or button_id == 'g2_t' then button_a = 'g2_b' button_b = 'g2_t'
-    elseif button_id == 'g2_r' or button_id == 'g2_g' then button_a = 'g2_r' button_b = 'g2_g'
-    elseif button_id == 'g2_o' or button_id == 'g2_y' then button_a = 'g2_o' button_b = 'g2_y'
-    end
-    if self.UI.getAttribute(button_id, 'outline') ~= color then
-        self.UI.setAttribute(button_a, 'outline', color)
-        self.UI.setAttribute(button_b, 'outline', color)
-    else
-        self.UI.setAttribute(button_a, 'outline', noColor)
-        self.UI.setAttribute(button_b, 'outline', noColor)
-    end
-end
-
-function g3(player, _, button_id)
-    local button_a = 'g3_p'
-    local button_b = 'g3_k'
-    local color = IDToColor[button_id:match('_(%l)')]
-    if button_id == 'g3_p' or button_id == 'g3_k' then button_a = 'g3_k' button_b = 'g3_p'
-    elseif button_id == 'g3_w' or button_id == 'g3_u' then button_a = 'g3_w' button_b = 'g3_u'
-    elseif button_id == 'g3_b' or button_id == 'g3_t' then button_a = 'g3_b' button_b = 'g3_t'
-    elseif button_id == 'g3_r' or button_id == 'g3_g' then button_a = 'g3_r' button_b = 'g3_g'
-    elseif button_id == 'g3_o' or button_id == 'g3_y' then button_a = 'g3_o' button_b = 'g3_y'
-    end
-    if self.UI.getAttribute(button_id, 'outline') ~= color then
-        self.UI.setAttribute(button_a, 'outline', color)
-        self.UI.setAttribute(button_b, 'outline', color)
-    else
-        self.UI.setAttribute(button_a, 'outline', noColor)
-        self.UI.setAttribute(button_b, 'outline', noColor)
-    end
-end
-
-function flip(player, _, button_id)
-    local color = IDToColor[button_id:match('_(%l)')]
-    local result = 'Tails'
-    if math.random(0, 100) > 50 then
-        result = 'Heads'
-    end
-    broadcastToAll(color .. " flipped " .. result .. '.', color)
-end
-
-function roll(player, _, button_id)
-    local color = IDToColor[button_id:match('_(%l)')]
-    local result = math.random(1, 6)
-    broadcastToAll(color .. " rolled a d6 and got " .. result .. '.', color)
-end
-
-function rd20(player, _, button_id)
-    local color = IDToColor[button_id:match('_(%l)')]
-    local result = math.random(1, 20)
-    broadcastToAll(color .. " rolled a d20 and got " .. result .. '.', color)
-end
-
-function mnrc(player, _, button_id)
-    local button_a = 'p'
-    local button_b = 'k'
-    local short = button_id:match('_(%l)')
-    local color = IDToColor[short]
-    if short == 'p' or short == 'k' then button_a = 'k' button_b = 'p'
-    elseif short == 'w' or short == 'u' then button_a = 'w' button_b = 'u'
-    elseif short == 'b' or short == 't' then button_a = 'b' button_b = 't'
-    elseif short == 'r' or short == 'g' then button_a = 'r' button_b = 'g'
-    elseif short == 'o' or short == 'y' then button_a = 'o' button_b = 'y'
-    end
-    if self.UI.getAttribute('btn_' .. short, 'outline') ~= color then
-        if self.UI.getAttribute('btn_' .. button_a, 'onClick') == 'mnrc' then
-            self.UI.setAttribute('btn_' .. button_a, 'outline', color)
-        end
-        if self.UI.getAttribute('btn_' .. button_b, 'onClick') == 'mnrc' then
-            self.UI.setAttribute('btn_' .. button_b, 'outline', color)
-        end
-        mnrcState[button_a] = color
-        mnrcState[button_b] = color
-    else
-        if self.UI.getAttribute('btn_' .. button_a, 'onClick') == 'mnrc' then
-            self.UI.setAttribute('btn_' .. button_a, 'outline', noColor)
-        end
-        if self.UI.getAttribute('btn_' .. button_b, 'onClick') == 'mnrc' then
-            self.UI.setAttribute('btn_' .. button_b, 'outline', noColor)
-        end
-        mnrcState[button_a] = noColor
-        mnrcState[button_b] = noColor
-    end
-end
-
----@param player any
----@param value '-1' | '-2'
----@param button_id string
-function dynt(player, value, button_id)
-    local button_a = 'p'
-    local button_b = 'k'
-    local short = button_id:match('_(%l)')
-    if short == 'p' or short == 'k' then button_a = 'k' button_b = 'p'
-    elseif short == 'w' or short == 'u' then button_a = 'w' button_b = 'u'
-    elseif short == 'b' or short == 't' then button_a = 'b' button_b = 't'
-    elseif short == 'r' or short == 'g' then button_a = 'r' button_b = 'g'
-    elseif short == 'o' or short == 'y' then button_a = 'o' button_b = 'y'
-    end
-    if dyntState[short] == nightColor then
-        if self.UI.getAttribute('btn_' .. button_a, 'onClick') == 'dynt' then
-            self.UI.setAttribute('btn_' .. button_a, 'outline', dayColor)
-        end
-        if self.UI.getAttribute('btn_' .. button_b, 'onClick') == 'dynt' then
-            self.UI.setAttribute('btn_' .. button_b, 'outline', dayColor)
-        end
-        dyntState[button_a] = dayColor
-        dyntState[button_b] = dayColor
-    elseif dyntState[short] == dayColor then
-        if self.UI.getAttribute('btn_' .. button_a, 'onClick') == 'dynt' then
-            self.UI.setAttribute('btn_' .. button_a, 'outline', nightColor)
-        end
-        if self.UI.getAttribute('btn_' .. button_b, 'onClick') == 'dynt' then
-            self.UI.setAttribute('btn_' .. button_b, 'outline', nightColor)
-        end
-        dyntState[button_a] = nightColor
-        dyntState[button_b] = nightColor
-    elseif value == '-1' then
-        if self.UI.getAttribute('btn_' .. button_a, 'onClick') == 'dynt' then
-            self.UI.setAttribute('btn_' .. button_a, 'outline', dayColor)
-        end
-        if self.UI.getAttribute('btn_' .. button_b, 'onClick') == 'dynt' then
-            self.UI.setAttribute('btn_' .. button_b, 'outline', dayColor)
-        end
-        dyntState[button_a] = dayColor
-        dyntState[button_b] = dayColor
-    elseif value == '-2' then
-        if self.UI.getAttribute('btn_' .. button_a, 'onClick') == 'dynt' then
-            self.UI.setAttribute('btn_' .. button_a, 'outline', noColor)
-        end
-        if self.UI.getAttribute('btn_' .. button_b, 'onClick') == 'dynt' then
-            self.UI.setAttribute('btn_' .. button_b, 'outline', noColor)
-        end
-        dyntState[button_a] = noColor
-        dyntState[button_b] = noColor
-    end
-end
-
-function swap(player, _, button_id)
-    local pId = button_id:match('_(%l)')
-    local dir = button_id:match('s(%u)_%l') -- N or P
-    local curr = self.UI.getAttribute('btn_' .. pId, 'onClick')
-    local nextFunc = {
-        ['roll'] = 'flip',
-        ['flip'] = 'rd20',
-        ['rd20'] = 'mnrc',
-        ['mnrc'] = 'dynt',
-        ['dynt'] = 'layt',
-        ['layt'] = 'roll'
-    }
-    local prevFunc = {
-        ['roll'] = 'layt',
-        ['flip'] = 'roll',
-        ['rd20'] = 'flip',
-        ['mnrc'] = 'rd20',
-        ['dynt'] = 'mnrc',
-        ['layt'] = 'dynt'
-    }
-    local func = prevFunc
-    if dir == 'N' then
-        func = nextFunc
-    end
-    self.UI.setAttribute('btn_' .. pId, 'onClick', func[curr])
-    self.UI.setAttribute('btn_' .. pId, 'icon', func[curr])
-    if func[curr] == 'mnrc' then
-        self.UI.setAttribute('btn_' .. pId, 'outline', mnrcState[pId])
-    elseif func[curr] == 'dynt' then
-        self.UI.setAttribute('btn_' .. pId, 'outline', dyntState[pId])
-    else
-        self.UI.setAttribute('btn_' .. pId, 'outline', noColor)
-    end
-end
-
-function layt(player, value, button_id)
-    local pId = button_id:match('_(%l)')
-    -- Hardcoded positions of library zones
-    local libraries = { p = { 64.02259, 4.377451, 13.24275 }, k = { 40.84833, 4.377451, -13.30888 }, w = { 12.9248, 4.377451, -13.32137 }, b = { -14.99426, 4.377451, -13.31943 }, r = { -42.91833, 4.377451, -13.3373 }, o = { -70.83037, 4.377451, -13.31797 }, y = { -47.63555, 4.377451, 13.2475 }, g = { -19.71588, 4.377451, 13.24803 }, t = { 8.197245, 4.377451, 13.26315 }, u = { 36.1164, 4.377451, 13.25679 } }
-    local deck = nil
-    local hits = Physics.cast({ origin = libraries[pId], type = 1, direction = { 0, 1, 0 }, max_distance = 3 })
-    for _, h in ipairs(hits) do
-        if h.hit_object.type == 'Deck' then
-            deck = h.hit_object
-        end
-    end
-    if deck == nil then
-        printToColor('Place a deck in your library slot.', player.color)
-        return
-    end
-    local pos_starting = { x = 0, y = 0, z = 0 }
-    local zone_guid = '5a1314'
-    if pId == 'k' then zone_guid = '6b2479' end
-    if pId == 'w' then zone_guid = 'ed8834' end
-    if pId == 'b' then zone_guid = '1451b7' end
-    if pId == 'r' then zone_guid = '2c271a' end
-    if pId == 'o' then zone_guid = '00a854' end
-    if pId == 'y' then zone_guid = '6c87b2' end
-    if pId == 'g' then zone_guid = 'fd020e' end
-    if pId == 't' then zone_guid = '90ea3b' end
-    if pId == 'u' then zone_guid = '6180e9' end
-    if pId == 'p' then zone_guid = '7058c4' end
-    local zn = getObjectFromGUID(zone_guid)
-    pos_starting = zn.getPosition()
-    pos_starting:setAt('x', pos_starting.x - 10.35)
-    if pId:match('p|u|t|g|y') then
-        pos_starting:setAt('z', pos_starting.z - 7)
-    else
-        pos_starting:setAt('z', pos_starting.z + 7)
-    end
-
-    -- Returns the first card alphabetically by name
-    function findNextCardIndex()
-        local orderList = {}
-        for _, card in ipairs(deck.getObjects()) do
-            if card.nickname ~= "" then
-                local insertTable = { name = card.nickname, index = card.index }
-                table.insert(orderList, insertTable)
-            end
-        end
-        -- Sort ordered list
-        local sort_func = function(a, b) return a["name"] > b["name"] end
-        table.sort(orderList, sort_func)
-        -- Add no-names onto start
-        for _, card in ipairs(deck.getObjects()) do
-            if card.nickname == "" then
-                local insertTable = { name = card.nickname, index = card.index }
-                table.insert(orderList, 1, insertTable)
-            end
-        end
-        return orderList[1].index
-    end
-
-    -- Placement
-    local face_down = deck.is_face_down
-    local wid = deck.getBoundsNormalized().size.x + 0.1
-    local hgt = deck.getBoundsNormalized().size.z + 0.1
-    local lastCard = nil
-    local rowStep, colStep = 0, 0
-    for _ = 1, 50 do
-        local pos_local = {
-            x = pos_starting.x + wid * colStep,
-            y = pos_starting.y,
-            z = pos_starting.z - hgt * rowStep,
-        }
-        colStep = colStep + 1
-        if colStep > 9 then
-            colStep = 0
-            rowStep = rowStep + 1
-        end
-        if lastCard == nil then
-            -- Handles most cards
-            local nextIndex = findNextCardIndex()
-            deck.takeObject({ position = pos_local, flip = face_down, index = nextIndex })
-            lastCard = deck.remainder
-        else
-            -- Handles the leftover card
-            lastCard.setPosition(pos_local)
-            if face_down then lastCard.flip() end
-        end
-        if deck == nil then break end
-    end
-end
+mod_name,version='Lavafume',1.09;gh_script,gh_ui='https://raw.githubusercontent.com/iammoonman/cera-roe/main/lavafume/main.lua','https://raw.githubusercontent.com/iammoonman/cera-roe/main/lavafume/ui.xml'IDToColor={['k']='Pink',['w']='White',['b']='Brown',['r']='Red',['o']='Orange',['y']='Yellow',['g']='Green',['t']='Teal',['u']='Blue',['p']='Purple'}mnrcState={['k']='',['w']='',['b']='',['r']='',['o']='',['y']='',['g']='',['t']='',['u']='',['p']=''}dyntState={['k']='',['w']='',['b']='',['r']='',['o']='',['y']='',['g']='',['t']='',['u']='',['p']=''}dayColor='Yellow'nightColor='Red'noColor=''function onLoad(a)self.interactable=false;WebRequest.get(gh_script,self,'GetFreshScript')if a~=''then local b=JSON.decode(a)mnrcState=b.mnrcState;dyntState=b.dyntState end end;function onSave()return JSON.encode({mnrcState=mnrcState,dyntState=dyntState})end;function GetFreshScript(c)if c==nil then return end;local d=c.text:match('mod_name, version = \'Lavafume\', (%d+%p%d+)')log('GITHUB Version '..d)if d then d=tonumber(d)else d=version end;if version<d then self.setLuaScript(c.text)self.reload()end;WebRequest.get(gh_ui,self,'GetFreshXML')end;function GetFreshXML(c)if c==nil then return end;self.UI.setXml(c.text)end;function ping(e,f,g)local h=IDToColor[g:match('_(%l)')]for i,j in ipairs(Player.getPlayers())do if j.host then j.broadcast(h..' needs assistance.')end end end;function untap(e,f,g)local h=IDToColor[g:match('_(%l)')]local k='5a1314'local l=270;if h=='Pink'then k='6b2479'l=180 end;if h=='White'then k='ed8834'l=180 end;if h=='Brown'then k='1451b7'l=180 end;if h=='Red'then k='2c271a'l=180 end;if h=='Orange'then k='00a854'l=180 end;if h=='Yellow'then k='6c87b2'l=0 end;if h=='Green'then k='fd020e'l=0 end;if h=='Teal'then k='90ea3b'l=0 end;if h=='Blue'then k='6180e9'l=0 end;if h=='Purple'then k='7058c4'l=0 end;local m=getObjectFromGUID(k)for f,n in ipairs(m.getObjects())do if n.type=="Card"then local o=n.getRotation()local p=o.z;n.setRotationSmooth({x=0,y=l,z=p})end end end;function g1(e,f,g)local q='g1_k'local r='g1_p'local h=IDToColor[g:match('_(%l)')]if g=='g1_p'or g=='g1_k'then q='g1_k'r='g1_p'elseif g=='g1_w'or g=='g1_u'then q='g1_w'r='g1_u'elseif g=='g1_b'or g=='g1_t'then q='g1_b'r='g1_t'elseif g=='g1_r'or g=='g1_g'then q='g1_r'r='g1_g'elseif g=='g1_o'or g=='g1_y'then q='g1_o'r='g1_y'end;if self.UI.getAttribute(g,'outline')~=h then self.UI.setAttribute(q,'outline',h)self.UI.setAttribute(r,'outline',h)else self.UI.setAttribute(q,'outline',noColor)self.UI.setAttribute(r,'outline',noColor)end end;function g2(e,f,g)local q='g2_p'local r='g2_k'local h=IDToColor[g:match('_(%l)')]if g=='g2_p'or g=='g2_k'then q='g2_k'r='g2_p'elseif g=='g2_w'or g=='g2_u'then q='g2_w'r='g2_u'elseif g=='g2_b'or g=='g2_t'then q='g2_b'r='g2_t'elseif g=='g2_r'or g=='g2_g'then q='g2_r'r='g2_g'elseif g=='g2_o'or g=='g2_y'then q='g2_o'r='g2_y'end;if self.UI.getAttribute(g,'outline')~=h then self.UI.setAttribute(q,'outline',h)self.UI.setAttribute(r,'outline',h)else self.UI.setAttribute(q,'outline',noColor)self.UI.setAttribute(r,'outline',noColor)end end;function g3(e,f,g)local q='g3_p'local r='g3_k'local h=IDToColor[g:match('_(%l)')]if g=='g3_p'or g=='g3_k'then q='g3_k'r='g3_p'elseif g=='g3_w'or g=='g3_u'then q='g3_w'r='g3_u'elseif g=='g3_b'or g=='g3_t'then q='g3_b'r='g3_t'elseif g=='g3_r'or g=='g3_g'then q='g3_r'r='g3_g'elseif g=='g3_o'or g=='g3_y'then q='g3_o'r='g3_y'end;if self.UI.getAttribute(g,'outline')~=h then self.UI.setAttribute(q,'outline',h)self.UI.setAttribute(r,'outline',h)else self.UI.setAttribute(q,'outline',noColor)self.UI.setAttribute(r,'outline',noColor)end end;function flip(e,f,g)local h=IDToColor[g:match('_(%l)')]local s='Tails'if math.random(0,100)>50 then s='Heads'end;broadcastToAll(h.." flipped "..s..'.',h)end;function roll(e,f,g)local h=IDToColor[g:match('_(%l)')]local s=math.random(1,6)broadcastToAll(h.." rolled a d6 and got "..s..'.',h)end;function rd20(e,f,g)local h=IDToColor[g:match('_(%l)')]local s=math.random(1,20)broadcastToAll(h.." rolled a d20 and got "..s..'.',h)end;function mnrc(e,f,g)local q='p'local r='k'local t=g:match('_(%l)')local h=IDToColor[t]if t=='p'or t=='k'then q='k'r='p'elseif t=='w'or t=='u'then q='w'r='u'elseif t=='b'or t=='t'then q='b'r='t'elseif t=='r'or t=='g'then q='r'r='g'elseif t=='o'or t=='y'then q='o'r='y'end;if self.UI.getAttribute('btn_'..t,'outline')~=h then if self.UI.getAttribute('btn_'..q,'onClick')=='mnrc'then self.UI.setAttribute('btn_'..q,'outline',h)end;if self.UI.getAttribute('btn_'..r,'onClick')=='mnrc'then self.UI.setAttribute('btn_'..r,'outline',h)end;mnrcState[q]=h;mnrcState[r]=h else if self.UI.getAttribute('btn_'..q,'onClick')=='mnrc'then self.UI.setAttribute('btn_'..q,'outline',noColor)end;if self.UI.getAttribute('btn_'..r,'onClick')=='mnrc'then self.UI.setAttribute('btn_'..r,'outline',noColor)end;mnrcState[q]=noColor;mnrcState[r]=noColor end end;function dynt(e,u,g)local q='p'local r='k'local t=g:match('_(%l)')if t=='p'or t=='k'then q='k'r='p'elseif t=='w'or t=='u'then q='w'r='u'elseif t=='b'or t=='t'then q='b'r='t'elseif t=='r'or t=='g'then q='r'r='g'elseif t=='o'or t=='y'then q='o'r='y'end;if dyntState[t]==nightColor then if self.UI.getAttribute('btn_'..q,'onClick')=='dynt'then self.UI.setAttribute('btn_'..q,'outline',dayColor)end;if self.UI.getAttribute('btn_'..r,'onClick')=='dynt'then self.UI.setAttribute('btn_'..r,'outline',dayColor)end;dyntState[q]=dayColor;dyntState[r]=dayColor elseif dyntState[t]==dayColor then if self.UI.getAttribute('btn_'..q,'onClick')=='dynt'then self.UI.setAttribute('btn_'..q,'outline',nightColor)end;if self.UI.getAttribute('btn_'..r,'onClick')=='dynt'then self.UI.setAttribute('btn_'..r,'outline',nightColor)end;dyntState[q]=nightColor;dyntState[r]=nightColor elseif u=='-1'then if self.UI.getAttribute('btn_'..q,'onClick')=='dynt'then self.UI.setAttribute('btn_'..q,'outline',dayColor)end;if self.UI.getAttribute('btn_'..r,'onClick')=='dynt'then self.UI.setAttribute('btn_'..r,'outline',dayColor)end;dyntState[q]=dayColor;dyntState[r]=dayColor elseif u=='-2'then if self.UI.getAttribute('btn_'..q,'onClick')=='dynt'then self.UI.setAttribute('btn_'..q,'outline',noColor)end;if self.UI.getAttribute('btn_'..r,'onClick')=='dynt'then self.UI.setAttribute('btn_'..r,'outline',noColor)end;dyntState[q]=noColor;dyntState[r]=noColor end end;function swap(e,f,g)local v=g:match('_(%l)')local w=g:match('s(%u)_%l')local x=self.UI.getAttribute('btn_'..v,'onClick')local y={['roll']='flip',['flip']='rd20',['rd20']='mnrc',['mnrc']='dynt',['dynt']='layt',['layt']='roll'}local z={['roll']='layt',['flip']='roll',['rd20']='flip',['mnrc']='rd20',['dynt']='mnrc',['layt']='dynt'}local A=z;if w=='N'then A=y end;self.UI.setAttribute('btn_'..v,'onClick',A[x])self.UI.setAttribute('btn_'..v,'icon',A[x])if A[x]=='mnrc'then self.UI.setAttribute('btn_'..v,'outline',mnrcState[v])elseif A[x]=='dynt'then self.UI.setAttribute('btn_'..v,'outline',dyntState[v])else self.UI.setAttribute('btn_'..v,'outline',noColor)end end;function layt(e,u,g)local v=g:match('_(%l)')local B={p={64.02259,4.377451,13.24275},k={40.84833,4.377451,-13.30888},w={12.9248,4.377451,-13.32137},b={-14.99426,4.377451,-13.31943},r={-42.91833,4.377451,-13.3373},o={-70.83037,4.377451,-13.31797},y={-47.63555,4.377451,13.2475},g={-19.71588,4.377451,13.24803},t={8.197245,4.377451,13.26315},u={36.1164,4.377451,13.25679}}local C=nil;local D=Physics.cast({origin=B[v],type=1,direction={0,1,0},max_distance=3})for f,E in ipairs(D)do if E.hit_object.type=='Deck'then C=E.hit_object end end;if C==nil then printToColor('Place a deck in your library slot.',e.color)return end;local F={x=0,y=0,z=0}local G='5a1314'if v=='k'then G='6b2479'end;if v=='w'then G='ed8834'end;if v=='b'then G='1451b7'end;if v=='r'then G='2c271a'end;if v=='o'then G='00a854'end;if v=='y'then G='6c87b2'end;if v=='g'then G='fd020e'end;if v=='t'then G='90ea3b'end;if v=='u'then G='6180e9'end;if v=='p'then G='7058c4'end;local H=getObjectFromGUID(G)F=H.getPosition()F:setAt('x',F.x-10.35)if v:match('p|u|t|g|y')then F:setAt('z',F.z-7)else F:setAt('z',F.z+7)end;function findNextCardIndex()local I={}for f,J in ipairs(C.getObjects())do if J.nickname~=""then local K={name=J.nickname,index=J.index}table.insert(I,K)end end;local L=function(M,N)return M["name"]>N["name"]end;table.sort(I,L)for f,J in ipairs(C.getObjects())do if J.nickname==""then local K={name=J.nickname,index=J.index}table.insert(I,1,K)end end;return I[1].index end;local O=C.is_face_down;local P=C.getBoundsNormalized().size.x+0.1;local Q=C.getBoundsNormalized().size.z+0.1;local R=nil;local S,T=0,0;for f=1,50 do local U={x=F.x+P*T,y=F.y,z=F.z-Q*S}T=T+1;if T>9 then T=0;S=S+1 end;if R~=nil then R.setPosition(U)if R.is_face_down then R.flip()end;break end;if R==nil then local V=findNextCardIndex()C.takeObject({position=U,flip=O,index=V})R=C.remainder end end end
