+++ lavafume/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "luamin": "^1.0.4"
+  }
+}
